int to float

+++ calculator.py
@@ -1,19 +1,13 @@
 def main():
-    x = int(input("what's x? "))
-    print("x cube is " , cube (x))
+    x = float(input("what's x? "))
+    print("x cube is " , cube(x))
 
 
 
-def  cube(n):
+def cube(n):
     return n**3
 
 main()
-
-
-
-
-
-
